Reader: use feature flag for site search dev
