+++ client/reader/search-stream/with-sites.jsx
@@ -1,0 +1,17 @@
+/**
+ * External Dependencies
+ */
+import React from 'react';
+import { localize } from 'i18n-calypso';
+
+/**
+ * Internal Dependencies
+ */
+
+class SearchStream extends React.Component {
+	render() {
+		return <p> shiny search page with site search functionality </p>;
+	}
+}
+
+export default localize( SearchStream );
